+++ src/Models/VerbStateEvent.php
@@ -7,8 +7,6 @@ use Thunk\Verbs\State;
 
 class VerbStateEvent extends Model
 {
-    use UsesVerbsIdType;
-
     public $table = 'verb_state_events';
 
     public $guarded = [];
@@ -22,4 +20,9 @@ class VerbStateEvent extends Model
     {
         return $this->state_type::load($this->state_id);
     }
+	
+	public function getIncrementing()
+	{
+		return false;
+	}
 }
